console font options change to particle sys so particles with short (i.e < 0.1s) longevity always have 1.0 alpha sprites3d can have XY scaling (for aspect ratio particles) some ui tweaks UIX dynamic text that uses callback to get current text to be displayed Util function SysFileAllocateAndLoad

+++ Pub/LibCode/CodeUtil/StringFunctions.cpp
@@ -85,7 +85,7 @@ void		EnsureFilenameHasExtension(char* acFilename, const char* acExtension)
 {
 	// Check if the filename already has the extension
 	int nExtLen = strlen(acExtension);
-	if (strlen(acFilename) >= nExtLen)
+	if ((int)strlen(acFilename) >= nExtLen)
 	{
 		if (stricmp(acExtension, &acFilename[strlen(acFilename) - nExtLen]) == 0)
 		{
